Revise config model (keep classes used for deserialization internal + add public interfaces to make data available to users)
