+++ src/ConfigCatClient/Hooks/Hooks.cs
@@ -88,7 +88,7 @@ internal class Hooks : IProvidesHooks
     }
 
     [MethodImpl(MethodImplOptions.NoInlining)]
-    internal void RaiseConfigChanged(object newConfig)
+    internal void RaiseConfigChanged(IConfig newConfig)
     {
         if (this.eventHandlers.ConfigChanged is { } configChanged && TryGetSender(out var client))
         {
