+++ src/ConfigCatClient/Models/RedirectMode.cs
@@ -1,0 +1,8 @@
+namespace ConfigCat.Client;
+
+internal enum RedirectMode : byte
+{
+    No = 0,
+    Should = 1,
+    Force = 2
+}
