+++ src/ConfigCatClient/Models/SettingType.cs
@@ -1,0 +1,10 @@
+namespace ConfigCat.Client;
+
+public enum SettingType : byte
+{
+    Boolean = 0,
+    String = 1,
+    Int = 2,
+    Double = 3,
+    Unknown = byte.MaxValue,
+}
